Create LF/HF function
Frequency domain features: 
Although the time domain parameters are effective, they do not have the ability of discrimination between the sympathetic and parasympathetic contents of the HRV signal. Highfrequency (HF) band (0.15-0.4 Hz) of HRV signal shows the cardiac vagal  activities such as Respiratory Sinus Arrhythmia (RSA). In fact, HF components are considered as the origin of parasympathetic activities of the cardiovascular system. On the other hand, the lowfrequency (LF) band (0.04-0.15 Hz) is related to the baroreceptor control and is mediated by symp athetic systems. 
In this function, the power spectral density (PSD) for the HF and LF bands are calculated and the ratio of the LF and HF bands power (LF/HF) is considered as the Frequency domain feature of the HRV signal.